Merge Develop into Master branch

+++ src/EA.Iws.Core/Notification/Overview/NotificationOverview.cs
@@ -36,7 +36,7 @@
 
         public SubmitSummaryData SubmitSummaryData { get; set; }
 
-        public int NotificationCharge { get; set; }
+        public decimal NotificationCharge { get; set; }
 
         public bool CanEditNotification { get; set; }
     }
